+++ src/api/chia/types/blockchain_format/program.ts
@@ -1,7 +1,5 @@
-import {bytes} from "../_python_types_";
+import {str} from "../_python_types_";
 
-export type SerializedProgram = {
-  _buf: bytes; // bytes = b""
-};
+export type SerializedProgram = str;
 
 export type Program = unknown;